+++ controllers/middleware/plugins.js
@@ -55,7 +55,7 @@ exports._M_.isAuthorized = [...exports._M_.isSignedIn, async (req, res, next) =>
     const [field, val] = req.params.canonical !== undefined
         ? ['canonical', req.params.canonical]
         : ['_id', _U_.string.readMongoId(req.url)];
-    if (await require('../../models/').postModel.count({ [field]: val, 'author._id': req.user }) !== 1) {               // tofix: find the post first then decide to give or not
+    if (await require('../../models/').postsModel.count({ [field]: val, 'author._id': req.user }) !== 1) {               // tofix: find the post first then decide to give or not
         throw new ClientError(20001);
     } else return next();
 }];