+++ controllers/routers/admin.js
@@ -7,7 +7,7 @@ module.exports = admin = {};
 // ==============================
 const { _M_ } = require('../middleware/plugins');
 const { _U_ } = require('../utilities/');
-const { configModel, mediaModel } = require('../../models/');
+const { configsModel, mediaModel } = require('../../models/');
 const { fetch } = require('../middleware/fetch');
 
 
@@ -24,7 +24,7 @@ admin.configs = {
         return next();
     },
     patch: async (req, res) => {
-        await configModel.updateSettings(req.body.configs);                                                             // tofix: pickup updated variables to avoid injections
+        await configsModel.updateSettings(req.body.configs);                                                             // tofix: pickup updated variables to avoid injections
         return res.redirect('back');
     },
 };
@@ -43,13 +43,11 @@ admin.upload = {   // todo: to be integrated in profile and media manager
 
 admin.stack = {
     get: (req, res, next) => {
-        switch (req.params['stackType'].toLowerCase()) {
-            case 'posts':
-                _M_.setTitleTag('Posts')(req, res);
-                return fetch({ num: 10 })(req, res, next);
-            default:
-                throw new _U_.error.HttpError(404);
-        }
+        const stackType = req.params['stackType'].toLowerCase();
+        if (['posts', 'media'].indexOf(stackType) !== -1) {
+            _M_.setTitleTag(stackType)(req, res);                                                                       // todo: capitalize
+            return fetch(`${stackType}Model`, { num: 10 })(req, res, next);
+        } else throw new _U_.error.HttpError(404);
     },
     patch: async (req, res) => res.redirect('back'),
 };