+++ controllers/routers/home.js
@@ -6,7 +6,7 @@ module.exports = home = {};
 //  DEPENDENCIES
 // ==============================
 const { _M_ } = require('../middleware/plugins');
-const { userModel } = require('../../models/');
+const { usersModel } = require('../../models/');
 
 
 
@@ -25,7 +25,7 @@ home.profile = {
     patch: async (req, res) => {
         const raw = { info: req.body.profile.info, nickname: req.body.profile.nickname };
         if (raw.info && raw.info.birthday) raw.info.birthday = new Date(raw.info.birthday);
-        await userModel.update({ _id: req.user._id }, { $set: raw });
+        await usersModel.update({ _id: req.user._id }, { $set: raw });
         req.flash('info', 'Your profile have been successfully updated!');
         return res.redirect('/home/profile');
     },