+++ controllers/routers/pages.js
@@ -1,0 +1,19 @@
+module.exports = pages = {};
+
+
+
+// ==============================
+//  DEPENDENCIES
+// ==============================
+const { _U_ } = require('../utilities/');
+const { postsModel } = require('../../models/');
+const { fetch } = require('../middleware/fetch');
+
+
+
+// ==============================
+//  CONTROLLERS
+// ==============================
+pages.search = {
+    get: fetch('postsModel'),
+};